Split relational query execution into focused modules
